setup different UIs for desktop and mobile
