+++ src/context/govPool/GovPoolProfileTabsContext/hooks/index.ts
@@ -1,4 +1,3 @@
-export { default as useAboutDao } from "./useAboutDao"
 export { default as useValidators } from "./useValidators"
 export { default as useDelegations } from "./useDelegations"
 export { default as useMyBalance } from "./useMyBalance"
