+++ src/pages/DaoProfile/tabs/index.ts
@@ -1,6 +1,2 @@
-export { default as DaoProfileTabAbout } from "./DaoProfileTabAbout"
-export { default as DaoProfileTabBalance } from "./DaoProfileTabBalance"
-export { default as DaoProfileTabValidators } from "./DaoProfileTabValidators"
-export { default as DaoProfileTabDelegations } from "./DaoProfileTabDelegations"
 export { default as TabFallback } from "./TabFallback"
-export { default as Routing } from "./Routing"
+export { default as Routing } from "./Routings"
