+++ src/pages/DaoProfile/tabs/mobile/index.ts
@@ -1,0 +1,4 @@
+export { default as DaoProfileTabAboutMobile } from "./DaoProfileTabAbout"
+export { default as DaoProfileTabBalanceMobile } from "./DaoProfileTabBalance"
+export { default as DaoProfileTabValidatorsMobile } from "./DaoProfileTabValidators"
+export { default as DaoProfileTabDelegationsMobile } from "./DaoProfileTabDelegations"
